# backend/openapi.yaml
openapi: 3.0.3
info:
  description: API para Sistema de Gestión de Condominios
  title: Casitas Teto API
  version: 1.0.0
servers:
- url: /
...
paths:
  /auth/login:
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/_auth_login_post_request"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
          description: Login exitoso
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: No autorizado
      security: []
      summary: Iniciar sesión
      tags:
      - Autenticación
...
  /condominios:
    get:
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_condominios_get_200_response"
          description: Lista de condominios
      summary: Listar condominios
      tags:
      - Condominios
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CondominioInput"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Condominio"
          description: Condominio creado
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Acceso denegado
      summary: Crear condominio (Solo Súper Administrador)
      tags:
      - Condominios
  /condominios/{condominioId}:
    delete:
      parameters:
      - description: ID del condominio
        in: path
        name: condominioId
        required: true
        schema:
          type: integer
      responses:
        "204":
          description: Condominio eliminado
      summary: Eliminar condominio
      tags:
      - Condominios
    get:
      parameters:
      - description: ID del condominio
        in: path
        name: condominioId
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Condominio"
          description: Detalles del condominio
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Recurso no encontrado
      summary: Obtener detalles de condominio
      tags:
      - Condominios
    put:
      parameters:
      - description: ID del condominio
        in: path
        name: condominioId
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CondominioInput"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Condominio"
          description: Condominio actualizado
      summary: Actualizar condominio
      tags:
      - Condominios
...
  /anuncios:
    get:
      parameters:
      - description: ID del condominio
        in: query
        name: condominioId
        required: false
        schema:
          type: integer
      - in: query
        name: activo
        schema:
          type: boolean
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/_anuncios_get_200_response"
          description: Lista de anuncios
      summary: Listar anuncios
      tags:
      - Anuncios
    post:
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnuncioInput"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Anuncio"
          description: Anuncio creado
      summary: Crear anuncio
      tags:
      - Anuncios
  /anuncios/{anuncioId}:
    delete:
      parameters:
      - description: ID del anuncio
        in: path
        name: anuncioId
        required: true
        schema:
          type: integer
      responses:
        "204":
          description: Anuncio eliminado
      summary: Eliminar anuncio
      tags:
      - Anuncios
    get:
      parameters:
      - description: ID del anuncio
        in: path
        name: anuncioId
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Anuncio"
          description: Detalles del anuncio
      summary: Obtener anuncio
      tags:
      - Anuncios
    put:
      parameters:
      - description: ID del anuncio
        in: path
        name: anuncioId
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnuncioInput"
        required: true
      responses:
        "200":
          description: Anuncio actualizado
      summary: Actualizar anuncio
      tags:
      - Anuncios
components:
...
  schemas:
    AuthResponse:
      example:
        usuario:
          apellido: apellido
          fechaCreacion: 2000-01-23T04:56:07.000+00:00
          condominioId: 6
          id: 0
          ultimoAcceso: 2000-01-23T04:56:07.000+00:00
          nombre: nombre
          email: email
          rol: SUPER_ADMINISTRADOR
          activo: true
        accessToken: accessToken
        refreshToken: refreshToken
      properties:
        accessToken:
          title: accessToken
          type: string
        refreshToken:
          title: refreshToken
          type: string
        usuario:
          $ref: "#/components/schemas/Usuario"
      title: AuthResponse
      type: object
    Usuario:
      example:
        apellido: apellido
        fechaCreacion: 2000-01-23T04:56:07.000+00:00
        condominioId: 6
        id: 0
        ultimoAcceso: 2000-01-23T04:56:07.000+00:00
        nombre: nombre
        email: email
        rol: SUPER_ADMINISTRADOR
        activo: true
      properties:
        id:
          title: id
          type: integer
        email:
          format: email
          title: email
          type: string
        nombre:
          title: nombre
          type: string
        apellido:
          title: apellido
          type: string
        rol:
          $ref: "#/components/schemas/RolUsuario"
        condominioId:
          nullable: true
          title: condominioId
          type: integer
        activo:
          title: activo
          type: boolean
        fechaCreacion:
          format: date-time
          title: fechaCreacion
          type: string
        ultimoAcceso:
          format: date-time
          nullable: true
          title: ultimoAcceso
          type: string
      title: Usuario
      type: object
...
    RolUsuario:
      enum:
      - SUPER_ADMINISTRADOR
      - ADMINISTRADOR
      - CONSERJE
      - DIRECTIVA
      - RESIDENTE
      title: RolUsuario
      type: string
    Condominio:
      example:
        direccion: direccion
        fechaCreacion: 2000-01-23T04:56:07.000+00:00
        totalViviendas: 6
        id: 0
        nombre: nombre
        activo: true
      properties:
        id:
          title: id
          type: integer
        nombre:
          title: nombre
          type: string
        direccion:
          title: direccion
          type: string
        totalViviendas:
          title: totalViviendas
          type: integer
        activo:
          title: activo
          type: boolean
        fechaCreacion:
          format: date-time
          title: fechaCreacion
          type: string
      title: Condominio
      type: object
    CondominioInput:
      example:
        direccion: direccion
        totalViviendas: 1
        nombre: nombre
      properties:
        nombre:
          title: nombre
          type: string
        direccion:
          title: direccion
          type: string
        totalViviendas:
          minimum: 1
          title: totalViviendas
          type: integer
      required:
      - direccion
      - nombre
      - totalViviendas
      title: CondominioInput
      type: object
...
    Anuncio:
      example:
        creadoPor: 1
        contenido: contenido
        titulo: titulo
        condominioId: 6
        id: 0
        fechaPublicacion: 2000-01-23T04:56:07.000+00:00
        activo: true
      properties:
        id:
          title: id
          type: integer
        condominioId:
          title: condominioId
          type: integer
        titulo:
          title: titulo
          type: string
        contenido:
          title: contenido
          type: string
        activo:
          title: activo
          type: boolean
        fechaPublicacion:
          format: date-time
          title: fechaPublicacion
          type: string
        creadoPor:
          title: creadoPor
          type: integer
      title: Anuncio
      type: object
    AnuncioInput:
      example:
        contenido: contenido
        titulo: titulo
        condominioId: 0
      properties:
        condominioId:
          title: condominioId
          type: integer
        titulo:
          title: titulo
          type: string
        contenido:
          title: contenido
          type: string
      required:
      - condominioId
      - contenido
      - tipo
      - titulo
      title: AnuncioInput
      type: object
    Error:
      example:
        codigo: codigo
        detalles:
        - detalles
        - detalles
        mensaje: mensaje
      properties:
        codigo:
          title: codigo
          type: string
        mensaje:
          title: mensaje
          type: string
        detalles:
          items:
            type: string
          title: detalles
          type: array
      title: Error
      type: object
...
    _auth_login_post_request:
      properties:
        email:
          format: email
          title: email
          type: string
        password:
          format: password
          title: password
          type: string
      required:
      - email
      - password
      title: _auth_login_post_request
      type: object
    _condominios_get_200_response:
      example:
        data:
        - direccion: direccion
          fechaCreacion: 2000-01-23T04:56:07.000+00:00
          totalViviendas: 6
          id: 0
          nombre: nombre
          activo: true
        - direccion: direccion
          fechaCreacion: 2000-01-23T04:56:07.000+00:00
          totalViviendas: 6
          id: 0
          nombre: nombre
          activo: true
      properties:
        data:
          items:
            $ref: "#/components/schemas/Condominio"
          title: data
          type: array
      title: _condominios_get_200_response
      type: object
...
    _anuncios_get_200_response:
      example:
        data:
        - creadoPor: 1
          contenido: contenido
          titulo: titulo
          condominioId: 6
          id: 0
          fechaPublicacion: 2000-01-23T04:56:07.000+00:00
          activo: true
        - creadoPor: 1
          contenido: contenido
          titulo: titulo
          condominioId: 6
          id: 0
          fechaPublicacion: 2000-01-23T04:56:07.000+00:00
          activo: true
      properties:
        data:
          items:
            $ref: "#/components/schemas/Anuncio"
          title: data
          type: array
      title: _anuncios_get_200_response
      type: object
...