# backend/src/app/models/anuncio.py
# coding: utf-8

"""
    Casitas Teto API

    API para Sistema de Gestión de Condominios

    The version of the OpenAPI document: 1.0.0
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501


from __future__ import annotations
import pprint
import re  # noqa: F401
import json




from datetime import datetime
from pydantic import BaseModel, ConfigDict, Field, StrictBool, StrictInt, StrictStr
from typing import Any, ClassVar, Dict, List, Optional
try:
    from typing import Self
except ImportError:
    from typing_extensions import Self

class Anuncio(BaseModel):
    """
    Anuncio
    """ # noqa: E501
    id: Optional[StrictInt] = None
    condominio_id: Optional[StrictInt] = Field(default=None, alias="condominioId")
    titulo: Optional[StrictStr] = None
    contenido: Optional[StrictStr] = None
    activo: Optional[StrictBool] = None
    fecha_publicacion: Optional[datetime] = Field(default=None, alias="fechaPublicacion")
    creado_por: Optional[StrictInt] = Field(default=None, alias="creadoPor")
    __properties: ClassVar[List[str]] = ["id", "condominioId", "titulo", "contenido", "activo", "fechaPublicacion", "creadoPor"]

    model_config = {
        "populate_by_name": True,
        "validate_assignment": True,
        "protected_namespaces": (),
    }


    def to_str(self) -> str:
        """Returns the string representation of the model using alias"""
        return pprint.pformat(self.model_dump(by_alias=True))

    def to_json(self) -> str:
        """Returns the JSON representation of the model using alias"""
        # TODO: pydantic v2: use .model_dump_json(by_alias=True, exclude_unset=True) instead
        return json.dumps(self.to_dict())

    @classmethod
    def from_json(cls, json_str: str) -> Self:
        """Create an instance of Anuncio from a JSON string"""
        return cls.from_dict(json.loads(json_str))

    def to_dict(self) -> Dict[str, Any]:
        """Return the dictionary representation of the model using alias.

        This has the following differences from calling pydantic's
        `self.model_dump(by_alias=True)`:

        * `None` is only added to the output dict for nullable fields that
          were set at model initialization. Other fields with value `None`
          are ignored.
        """
        _dict = self.model_dump(
            by_alias=True,
            exclude={
            },
            exclude_none=True,
        )
        return _dict

    @classmethod
    def from_dict(cls, obj: Dict) -> Self:
        """Create an instance of Anuncio from a dict"""
        if obj is None:
            return None

        if not isinstance(obj, dict):
            return cls.model_validate(obj)

        _obj = cls.model_validate({
            "id": obj.get("id"),
            "condominioId": obj.get("condominioId"),
            "titulo": obj.get("titulo"),
            "contenido": obj.get("contenido"),
            "activo": obj.get("activo"),
            "fechaPublicacion": obj.get("fechaPublicacion"),
            "creadoPor": obj.get("creadoPor")
        })
        return _obj