# backend/src/app/models/rol_usuario.py
# coding: utf-8

"""
    Casitas Teto API

    API para Sistema de Gestión de Condominios

    The version of the OpenAPI document: 1.0.0
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501


from __future__ import annotations
import json
import pprint
import re  # noqa: F401
from enum import Enum



try:
    from typing import Self
except ImportError:
    from typing_extensions import Self


class RolUsuario(str, Enum):
    """
    RolUsuario
    """

    """
    allowed enum values
    """
    SUPER_ADMINISTRADOR = 'SUPER_ADMINISTRADOR'
    ADMINISTRADOR = 'ADMINISTRADOR'
    CONSERJE = 'CONSERJE'
    DIRECTIVA = 'DIRECTIVA'
    RESIDENTE = 'RESIDENTE'

    @classmethod
    def from_json(cls, json_str: str) -> Self:
        """Create an instance of RolUsuario from a JSON string"""
        return cls(json.loads(json_str))